Put each model in a separate python file
